Added git usage section to readme
To maintain a consistent format to commits, and to ensure that I dont
forget about how to use git, the README has been updated to include a
section on how to push to this repo. This has been done to maintain
consistency and to ensure no silly mistakes get pushed to the master
branch (ive done this too much at this point).
+ added `Git Usage` section to layout common git etiquette
  + added common format for commit messages to ensure consistency
  + added branching quick-start to ensure isolation between features
  + added note about pull requests. needs discussion at next tutorial